+++ backend/src/baserow/contrib/database/models.py
@@ -3,7 +3,8 @@ from baserow.core.models import Application
 from .table.models import Table
 from .views.models import View, GridView, GridViewFieldOptions, ViewFilter
 from .fields.models import (
-    Field, TextField, NumberField, LongTextField, BooleanField, DateField, LinkRowField
+    Field, TextField, NumberField, LongTextField, BooleanField, DateField, LinkRowField,
+    URLField, EmailField
 )
 
 __all__ = [
@@ -11,7 +12,7 @@ __all__ = [
     'Table',
     'View', 'GridView', 'GridViewFieldOptions', 'ViewFilter',
     'Field', 'TextField', 'NumberField', 'LongTextField', 'BooleanField', 'DateField',
-    'LinkRowField'
+    'LinkRowField', 'URLField', 'EmailField'
 ]