+++ backend/src/baserow/contrib/database/views/view_filters.py
@@ -10,7 +10,7 @@ from django.db.models.fields.related import ManyToManyField
 
 from baserow.contrib.database.fields.field_types import (
     TextFieldType, LongTextFieldType, URLFieldType, NumberFieldType, DateFieldType,
-    LinkRowFieldType, BooleanFieldType
+    LinkRowFieldType, BooleanFieldType, EmailFieldType
 )
 
 from .registries import ViewFilterType
@@ -35,7 +35,8 @@ class EqualViewFilterType(ViewFilterType):
         LongTextFieldType.type,
         URLFieldType.type,
         NumberFieldType.type,
-        BooleanFieldType.type
+        BooleanFieldType.type,
+        EmailFieldType.type
     ]
 
     def get_filter(self, field_name, value, model_field):
@@ -70,6 +71,7 @@ class ContainsViewFilterType(ViewFilterType):
         TextFieldType.type,
         LongTextFieldType.type,
         URLFieldType.type,
+        EmailFieldType.type
     ]
 
     def get_filter(self, field_name, value, model_field):
@@ -248,7 +250,8 @@ class EmptyViewFilterType(ViewFilterType):
         NumberFieldType.type,
         BooleanFieldType.type,
         DateFieldType.type,
-        LinkRowFieldType.type
+        LinkRowFieldType.type,
+        EmailFieldType.type
     ]
 
     def get_filter(self, field_name, value, model_field):