+++ web-frontend/modules/database/plugin.js
@@ -4,6 +4,7 @@ import {
   TextFieldType,
   LongTextFieldType,
   URLFieldType,
+  EmailFieldType,
   LinkRowFieldType,
   NumberFieldType,
   BooleanFieldType,
@@ -54,4 +55,5 @@ export default ({ store, app }) => {
   app.$registry.register('field', new BooleanFieldType())
   app.$registry.register('field', new DateFieldType())
   app.$registry.register('field', new URLFieldType())
+  app.$registry.register('field', new EmailFieldType())
 }